Refactor uniswap types
