+++ src/common/interfaces/uniswap.interface.ts
@@ -7,3 +7,29 @@ export interface Network {
   isNew?: boolean;
   subgraphEndpoint: string;
 }
+
+export interface Tick {
+  tickIdx: string;
+  liquidityNet: string;
+  price0: string;
+  price1: string;
+}
+
+export interface Token {
+  id: string;
+  name: string;
+  symbol: string;
+  volumeUSD: string;
+  logoURI: string;
+  decimals: string;
+}
+
+export interface Pool {
+  id: string;
+  feeTier: string;
+  liquidity: string;
+  tick: string;
+  sqrtPrice: string;
+  token0Price: string;
+  token1Price: string;
+}
